feat(symgo): implement initial version of symgotest library
This commit introduces `symgotest`, a new testing library for the `symgo` symbolic execution engine, as outlined in `docs/plan-symgotest.md`.

The primary goal of `symgotest` is to provide a debugging-first testing experience, reducing boilerplate and making it easier to diagnose failures in `symgo`-based tools.

Key features in this initial implementation include:

- A core `symgotest.Run` function that automates test setup, execution, and result collection.
- Step-limiting (`WithMaxSteps`) to prevent infinite execution loops. This was implemented by adding a step counter to the `symgo` evaluator.
- An execution tracer that captures the evaluation flow and can be used to report detailed diagnostics on test failure.
- Convenience wrappers `RunExpression` and `RunStatements` for quickly testing small code snippets.
- Generic, type-safe assertion helpers (`AssertAs`, `AssertEqual`) to simplify test code.

To support these features, the `symgo` engine was enhanced:
- The `symgo.evaluator` now accepts a `WithMaxSteps` functional option.
- The `object.Tracer` interface was enhanced to pass a `TraceEvent` struct, providing more context to tracers.

A comprehensive test suite for `symgotest` has been added, covering both successful execution and failure cases. Documentation (`README.md`, `TODO.md`, `plan-symgotest.md`) has been added and updated to reflect these new features.

+++ symgotest/symgotest_test.go
@@ -88,7 +88,7 @@ func TestRunExpression(t *testing.T) {
 		if r.Error != nil {
 			t.Fatalf("Execution failed: %v", r.Error)
 		}
-		AssertEqual(t, r.ReturnValue, int64(3))
+		AssertEqual(t, r.ReturnValue, 3)
 	}
 	RunExpression(t, "1 + 2", action)
 }
@@ -104,7 +104,7 @@ func TestRunStatements(t *testing.T) {
 		}
 
 		variable := AssertAs[*object.Variable](t, val)
-		AssertEqual(t, variable.Value, int64(10))
+		AssertEqual(t, variable.Value, 10)
 	}
 	RunStatements(t, "x := 10", action)
 }
